framing: Phase A fixes — DINO text format, MPS dtypes, threshold tuning
DINO text format: transformers 5.x rejects [[prompt]] nesting for the
period-separated multi-class form; pass the prompt as a flat string.

MPS dtype: F16 trips a Metal kernel assertion on M5 Max
(MPSNDArrayMatrixMultiplication accumulator dtype mismatch). Switched
all model wrappers and pipeline default to F32. ~2x slower than F16
but reliable; for 360-image V1 runs that's ~2 min vs ~1 min — fine.

InsectSAM input tensors: the SAM processor produces float64 inputs by
default (Python float → torch default). MPS rejects float64, so cast
explicitly to F32 before .to(device) in both segmenter codepaths
(image embedding + bbox prompt).

BOX_THRESHOLD: lowered 0.25 → 0.15 so we capture bbox data even on
low-confidence detections. Anything below 0.40 conf still ends up
classified as framing_quality='hidden' per the classify rules, so
this just gives us richer metric data for those cases.

Phase A 10-point sanity gate: ALL GATES PASSED on 5-image probe.
Per-image latency: DINO ~280ms + InsectSAM ~25ms = ~305ms total.
360-image V1 pass projected at ~2 minutes.

+++ scripts/detect_subjects/pipeline.py
@@ -64,7 +64,7 @@ def run_v1_on_sample(
     gt_index: GroundTruthIndex | None = None,
     parquet_path: Path = PARQUET_PATH,
     device: str = "mps",
-    dtype: torch.dtype = torch.float16,
+    dtype: torch.dtype = torch.float32,
 ) -> dict:
     """Run V1 (DINO + InsectSAM) over every row in sample_rows."""
     completed = load_completed_pairs(parquet_path)
